# Define directly