--- a/lib/patches/sql_patches.rb
+++ b/lib/patches/sql_patches.rb
@@ -11,20 +11,8 @@ class SqlPatches
     @patched = val
   end
 
-  def self.class_exists?(name)
-    eval(name + ".class").to_s.eql?('Class')
-  rescue NameError
-    false
-  end
-
   def self.correct_version?(required_version, klass)
     Gem::Dependency.new('', required_version).match?('', klass::VERSION)
-  rescue NameError
-    false
-  end
-
-  def self.module_exists?(name)
-    eval(name + ".class").to_s.eql?('Module')
   rescue NameError
     false
   end
@@ -46,15 +34,15 @@ class SqlPatches
   end
 end
 
-require 'patches/db/mysql2'           if defined?(Mysql2::Client) && SqlPatches.class_exists?("Mysql2::Client")
-require 'patches/db/pg'               if defined?(PG::Result) && SqlPatches.class_exists?("PG::Result")
-require 'patches/db/oracle_enhanced'  if defined?(ActiveRecord::ConnectionAdapters::OracleEnhancedAdapter) && SqlPatches.class_exists?("ActiveRecord::ConnectionAdapters::OracleEnhancedAdapter") && SqlPatches.correct_version?('~> 1.5.0', ActiveRecord::ConnectionAdapters::OracleEnhancedAdapter)
-require 'patches/db/mongo'            if defined?(Mongo) && SqlPatches.module_exists?("Mongo")
-require 'patches/db/moped'            if defined?(Moped::Node) && SqlPatches.class_exists?("Moped::Node")
-require 'patches/db/plucky'           if defined?(Plucky::Query) && SqlPatches.class_exists?("Plucky::Query")
-require 'patches/db/rsolr'            if defined?(RSolr::Connection) && SqlPatches.class_exists?("RSolr::Connection") && RSolr::VERSION[0] != "0"
-require 'patches/db/sequel'           if defined?(Sequel::Database) && SqlPatches.unpatched? && SqlPatches.class_exists?("Sequel::Database")
-require 'patches/db/activerecord'     if defined?(ActiveRecord) && SqlPatches.unpatched? && SqlPatches.module_exists?("ActiveRecord")
-require 'patches/db/nobrainer'        if defined?(NoBrainer) && SqlPatches.module_exists?("NoBrainer")
-require 'patches/db/riak'             if defined?(Riak) && SqlPatches.module_exists?("Riak")
-require 'patches/db/neo4j'            if defined?(Neo4j::Core) && SqlPatches.class_exists?("Neo4j::Core::Query")
+require 'patches/db/mysql2'           if defined?(Mysql2::Client) && Mysql2::Client.class == Class
+require 'patches/db/pg'               if defined?(PG::Result) && PG::Result.class == Class
+require 'patches/db/oracle_enhanced'  if defined?(ActiveRecord::ConnectionAdapters::OracleEnhancedAdapter) && ActiveRecord::ConnectionAdapters::OracleEnhancedAdapter.class == Class && SqlPatches.correct_version?('~> 1.5.0', ActiveRecord::ConnectionAdapters::OracleEnhancedAdapter)
+require 'patches/db/mongo'            if defined?(Mongo) && Mongo.class == Module
+require 'patches/db/moped'            if defined?(Moped::Node) && Moped::Node.class == Class
+require 'patches/db/plucky'           if defined?(Plucky::Query) && Plucky::Query.class == Class
+require 'patches/db/rsolr'            if defined?(RSolr::Connection) && RSolr::Connection.class == Class && RSolr::VERSION[0] != "0"
+require 'patches/db/sequel'           if SqlPatches.unpatched? && defined?(Sequel::Database) && Sequel::Database.class == Class
+require 'patches/db/activerecord'     if SqlPatches.unpatched? && defined?(ActiveRecord) && ActiveRecord.class == Module
+require 'patches/db/nobrainer'        if defined?(NoBrainer) && NoBrainer.class == Module
+require 'patches/db/riak'             if defined?(Riak) && Riak.class == Module
+require 'patches/db/neo4j'            if defined?(Neo4j::Core) && Neo4j::Core::Query.class == ClassQuery