--- a/spec/components/sql_patches_spec.rb
+++ b/spec/components/sql_patches_spec.rb
@@ -28,32 +28,4 @@ describe SqlPatches do
       expect(SqlPatches).to be_unpatched
     end
   end
-
-  describe ".class_exists?(name)" do
-    it "detects non-existant" do
-      expect(SqlPatches.class_exists?("SomeRandomClassThatDoesntExist")).to eq(false)
-    end
-
-    it "detects module" do
-      expect(SqlPatches.class_exists?("Rack")).to eq(false)
-    end
-
-    it "detects class" do
-      expect(SqlPatches.class_exists?("SqlPatches")).to eq(true)
-    end
-  end
-
-  describe ".module_exists?(name)" do
-    it "detects non-existant" do
-      expect(SqlPatches.module_exists?("SomeRandomClassThatDoesntExist")).to eq(false)
-    end
-
-    it "detects module" do
-      expect(SqlPatches.module_exists?("Rack")).to eq(true)
-    end
-
-    it "detects class" do
-      expect(SqlPatches.module_exists?("SqlPatches")).to eq(false)
-    end
-  end
 end