@SamSaffron 2016-04-22
less evals are better totally happy with that, we could possibly extract a method here, but will leave it like this for now.